ai-services/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.6.0
motor==3.3.2
openai==1.3.7
transformers==4.36.0
torch==2.1.1
torchvision==0.16.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
pillow==10.1.0
pytesseract==0.3.10
spacy==3.7.2
nltk==3.8.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.25.2
jinja2==3.1.2
markdown==3.5.1
beautifulsoup4==4.12.2
requests==2.31.0
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pdfplumber==0.10.3
PyMuPDF==1.23.8
